refactor: Refactor admin screens and add admin-only access
This commit refactors the admin screens by breaking down the `AdminContentScreen` into smaller, more manageable widgets: `AdminChallengeSelectionScreen` and `CurrentChallengeAdminScreen`. This improves the maintainability of the admin-related code.

Additionally, this commit introduces a check in the `AdminScreen` to ensure that only non-anonymous users can access the admin features. This enhances the security of the admin panel.

Finally, the `CurrentChallengeMixin` and `CurrentChallengerMixin` have been updated to better handle null and admin users.